pyproject.toml
[tool.poetry]
name = "zelda"
version = "1.0.0"
description = "A GPT-4 powered application for summarization and note generation for Academic papers provided via uploading a PDF of the paper."
authors = ["Matthew Sullivan <matthew@mseeks.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20221105"
nltk = "^3.8.1"
flask = "^2.3.2"
boto3 = "^1.26.134"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
watchdog = "^3.0.0"
boto3-stubs = "^1.26.134"
mypy = "^1.3.0"
mypy-boto3-s3 = "^1.26.127"
types-botocore = "^1.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
